Deploy the db to mysql (resolve problem with XAMPP)

+++ api_servis/db/migrate/3_create_proizvodi.rb
@@ -8,7 +8,7 @@ class CreateProizvodi < ActiveRecord::Migration[6.1]
       t.integer :cena, null: false
 
       # foreign keys
-      t.integer :kategorija_id, null: false
+      t.bigint :kategorija_id, null: false
 
     end
     # add the foreign key constraint
